move github email request in a specific method
