+++ fastapi_sso/sso/github.py
@@ -1,6 +1,8 @@
 """Github SSO Oauth Helper class"""
 
 from fastapi_sso.sso.base import DiscoveryDocument, OpenID, SSOBase, SSOLoginError 
+from typing import Dict, Optional
+from httpx import AsyncClient
 
 
 class GithubSSO(SSOBase):
@@ -18,15 +20,33 @@ class GithubSSO(SSOBase):
             "emails_endpoint": "https://api.github.com/user/emails",
         }
 
+    @property
+    async def emails_endpoint(self) -> Optional[str]:
+        """Return `emails_endpoint` from discovery document"""
+        discovery = await self.get_discovery_document()
+        return discovery.get("emails_endpoint")
 
-    @classmethod
-    async def openid_from_response(cls, response: dict) -> OpenID:
-
+    async def get_extra_data(self, content: Dict, session: AsyncClient) -> Dict:
         # if the email is empty then it means that the user has set it has private
         # so the email must be retrieved using emails_endopoint.
         # This endpoint return the list of all the email addresses.
+        if content["email"] is not None:
+            return content
+
+        uri, headers, _ = self.oauth_client.add_token(await self.emails_endpoint)
+        email_response = await session.get(uri, headers=headers)
+        emails = email_response.json()
+        for email in emails:
+            if email["primary"]:
+                content["email"] = email["email"]
+                break
+
+        return content
+
+    @classmethod
+    async def openid_from_response(cls, response: dict) -> OpenID:
         if response["email"] is None:
-            raise SSOLoginError(401, "Failed to process login via GitHub, please set your email as public in GitHub settings")
+            raise SSOLoginError(401, "Failed to process login via GitHub")
 
         return OpenID(
             email=response["email"],
